Boston Dataset

A data frame reporting housing values in suburbs of Boston.

Usage

Boston

Format

A data frame with 506 rows and 14 columns

crim

numeric. Per capita crime rate by town.

zn

numeric. Proportion of residential land zoned for lots over 25,000 sq.ft.

indus

numeric. Proportion of non-retail business acres per town.

chas

integer. Charles River dummy variable (=1 if tract bounds river; 0 otherwise).

nox

numeric. Nitrogen oxides concentration (parts per 10 million).

rm

numeric. Average number of rooms per dwelling.

age

numeric. Proportion of owner-occupied units built prior to 1940.

dis

numeric. Weighted mean of distances to five Boston employment centers.

rad

integer. Index of accessibility to radial highways.

tax

numeric. full-value property-tax rate per $10,000.

ptratio

numeric. pupil-teacher ratio by town.

black

numeric. 1000(Bk - 0.63)^2 where Bk is the proportion of blacks by town

lstat

numeric. Lower status of the population (percent)

medv

numeric. Median value of owner-occupied homes in $1000s

Source

This data is loaded from MASS package in R