print.ryx

The print function takes correlations created by the ryx function and prints a table of correlations that notes when the p value is < 2e-16

# S3 method for ryx
print(x, digits = 3)

Examples

library(MASS)
x <- ryx(Boston, y="medv")
x
#>       variable          r        p sigif
#> cor10    lstat -0.7376627 <2.0e-16   ***
#> cor4        rm  0.6953599 <2.0e-16   ***
#> cor8   ptratio -0.5077867 <2.0e-16   ***
#> cor2     indus -0.4837252 <2.0e-16   ***
#> cor7       tax -0.4685359 <2.0e-16   ***
#> cor3       nox -0.4273208 <2.0e-16   ***
#> cor       crim -0.3883046 <2.0e-16   ***
#> cor5       age -0.3769546 <2.0e-16   ***
#> cor1        zn  0.3604453 <2.0e-16   ***
#> cor9     black  0.3334608 1.32e-14   ***
#> cor6       dis  0.2499287 1.21e-08   ***