Test correlations

Test the correlations between the response variable and all the other explanatory variables provided.

ryx(data, y, x)

Arguments

data

the dataset to test on

y

the name of the quantitative response variable (str)

x

a vector of names of quantitative explanatory variables. If not provided, then automatically use all the quantitative variables.

...

optional parameters passed to import functions

Value

y

The name of response variable

x

A vector of names of explanatory variables

df

A dataframe of correlations

Examples

# x is missing
ryx(mtcars, 'mpg')
#> Correlations of mpg with
#>  variable      r        p sigif
#>        wt -0.868 1.29e-10   ***
#>       cyl -0.852 6.11e-10   ***
#>      disp -0.848 9.38e-10   ***
#>        hp -0.776 1.79e-07   ***
#>      drat  0.681 1.78e-05   ***
#>        vs  0.664 3.42e-05   ***
#>        am  0.600 2.85e-04   ***
#>      carb -0.551 1.08e-03    **
#>      gear  0.480 5.40e-03    **
#>      qsec  0.419 1.71e-02     *

# give x
ryx(mtcars, 'mpg', c('hp', 'wt'))
#> Correlations of mpg with
#>  variable      r        p sigif
#>        wt -0.868 1.29e-10   ***
#>        hp -0.776 1.79e-07   ***