Prints a summary of the correlations between the selected variable and other variables in the dataset

Prints a summary of the correlations between the selected variable and other variables in the dataset

# S3 method for class 'ryx'
summary(x)

Arguments

x

An object of class ryx, generated by the ryx function. It contains the correlation, p-values, and significance levels between the selected variable and the other variables in the selected dataset.

Value

The function returns a printed summary including:

The variables being compared,

The median absolute correlation between the selected variable and the other variables,

The range of the correlations,

The number of variables that are significant at the p < 0.05 level.

Examples

# Example usage:
library(MASS)
object <- ryx(Boston, y="medv")
summary(object)
#> Correlating medv with crim zn indus nox rm age dis tax ptratio black lstat 
#> The median absolute correlation was 0.427 with a range from 0.25 to 0.738 
#> 11 out of 11 variables were significant at the p < 0.05 level.
data(mtcars)
object2 <- ryx(mtcars, y="mpg")
summary(object2)
#> Correlating mpg with cyl disp hp drat wt qsec vs am gear carb 
#> The median absolute correlation was 0.673 with a range from 0.419 to 0.868 
#> 10 out of 10 variables were significant at the p < 0.05 level.