Correlation Table

Creates a correlation table using numeric variables.

Creates a correlation table using numeric variables.

Usage

# S3 method for class 'ryx'
print(x, ...)

# S3 method for class 'ryx'
print(x, ...)

Arguments

x

string, a list of numeric variables. Defaults to all variables in the dataset provided.

data

a data frame

y

a numeric variable

Value

method used, dataset used, variables used for outliers detected, indices of any detected outliers, scores for the outliers, and values for optional parameters

dataframe with correlations of y variable with all x-variables, p-value, and level of significance for each x variable.

Examples

multiOutliers(mtcarsOutliers, method="mahalanobis", alpha=0.1)
#> Error in multiOutliers(mtcarsOutliers, method = "mahalanobis", alpha = 0.1): could not find function "multiOutliers"
multiOutliers(mtcarsOutliers, method="LoF", minPts=5)
#> Error in multiOutliers(mtcarsOutliers, method = "LoF", minPts = 5): could not find function "multiOutliers"
multiOutliers(mtcarsOutliers, method="kNN", k=5, threshold=.95)
#> Error in multiOutliers(mtcarsOutliers, method = "kNN", k = 5, threshold = 0.95): could not find function "multiOutliers"
multiOutliers(mtcarsOutliers, method="iForest", ntrees = 50)
#> Error in multiOutliers(mtcarsOutliers, method = "iForest", ntrees = 50): could not find function "multiOutliers"
library(MASS)
x <- ryx(Boston, y = "medv")
x
#> Correlations of medv with
#>  variable          r            p sigif
#>     lstat -0.7376627 5.081103e-88   ***
#>        rm  0.6953599 2.487229e-74   ***
#>   ptratio -0.5077867 1.609509e-34   ***
#>     indus -0.4837252 4.900260e-31   ***
#>       tax -0.4685359 5.637734e-29   ***
#>       nox -0.4273208 7.065042e-24   ***
#>      crim -0.3883046 1.173987e-19   ***
#>       age -0.3769546 1.569982e-18   ***
#>        zn  0.3604453 5.713584e-17   ***
#>     black  0.3334608 1.318113e-14   ***
#>       dis  0.2499287 1.206612e-08   ***
x <- ryx(iris, y = "Sepal.Length")
x
#> Correlations of Sepal.Length with
#>      variable          r            p sigif
#>  Petal.Length  0.8717538 1.038667e-47   ***
#>   Petal.Width  0.8179411 2.325498e-37   ***
#>   Sepal.Width -0.1175698 1.518983e-01