Printing correlations between selected variable and other variables in selected dataset

Prints correlations between selected variable and other variables in selected dataset

# S3 method for class 'ryx'
print(x)

Arguments

x

An object of class ryx, generated by the ryx function. It contains the correlation, p-values, and level of significance between the selected variable and specified variables in the selected dataset. If none are specified, the selected variable will be compared to all other numeric variables in the dataset, excluding itself.

Value

The function returns a formatted print of the correlations between the selected variable and all the other variables in the selected dataset.

Examples

# Example usage:
library(MASS)
object <- ryx(Boston, y="medv")
print(object)
#> Correlations of medv with variables: crim, zn, indus, nox, rm, age, dis, tax, ptratio, black, lstat 
#>       variable          r            p sigif
#> cor10    lstat -0.7376627 5.081103e-88   ***
#> cor4        rm  0.6953599 2.487229e-74   ***
#> cor8   ptratio -0.5077867 1.609509e-34   ***
#> cor2     indus -0.4837252 4.900260e-31   ***
#> cor7       tax -0.4685359 5.637734e-29   ***
#> cor3       nox -0.4273208 7.065042e-24   ***
#> cor       crim -0.3883046 1.173987e-19   ***
#> cor5       age -0.3769546 1.569982e-18   ***
#> cor1        zn  0.3604453 5.713584e-17   ***
#> cor9     black  0.3334608 1.318113e-14   ***
#> cor6       dis  0.2499287 1.206612e-08   ***
# Example that compares selected variable with specified variables from a given dataset
data(mtcars)
object2 <- ryx(mtcars, y = "mpg", x = c("qsec","gear","carb"))
print(object2)
#> Correlations of mpg with variables: qsec, gear, carb 
#>      variable          r           p sigif
#> cor2     carb -0.5509251 0.001084446    **
#> cor1     gear  0.4802848 0.005400948    **
#> cor      qsec  0.4186840 0.017081988     *