Printing correlations between selected variable and other variables in selected dataset

Prints correlations between selected variable and other variables in selected dataset

Usage

# S3 method for class 'ryx'
print(x)

Arguments

x

An object of class ryx, generated by the ryx function. It contains the correlation, p-values, and level of significance between the selected variable and all the other variables in the selected dataset.

Value

The function returns a formatted print of the correlations between the selected variable and all the other variables in the selected dataset.

Examples

# Example usage:
library(MASS)
object <- ryx(Boston, y="medv")
print(object)
#> Correlations of medv with variables: crim, zn, indus, nox, rm, age, dis, tax, ptratio, black, lstat 
#>       variable          r            p sigif
#> cor10    lstat -0.7376627 5.081103e-88   ***
#> cor4        rm  0.6953599 2.487229e-74   ***
#> cor8   ptratio -0.5077867 1.609509e-34   ***
#> cor2     indus -0.4837252 4.900260e-31   ***
#> cor7       tax -0.4685359 5.637734e-29   ***
#> cor3       nox -0.4273208 7.065042e-24   ***
#> cor       crim -0.3883046 1.173987e-19   ***
#> cor5       age -0.3769546 1.569982e-18   ***
#> cor1        zn  0.3604453 5.713584e-17   ***
#> cor9     black  0.3334608 1.318113e-14   ***
#> cor6       dis  0.2499287 1.206612e-08   ***
data(mtcars)
object2 <- ryx(mtcars, y="mpg")
print(object2)
#> Correlations of mpg with variables: cyl, disp, hp, drat, wt, qsec, vs, am, gear, carb 
#>      variable          r            p sigif
#> cor4       wt -0.8676594 1.293959e-10   ***
#> cor       cyl -0.8521620 6.112687e-10   ***
#> cor1     disp -0.8475514 9.380327e-10   ***
#> cor2       hp -0.7761684 1.787835e-07   ***
#> cor3     drat  0.6811719 1.776240e-05   ***
#> cor6       vs  0.6640389 3.415937e-05   ***
#> cor7       am  0.5998324 2.850207e-04   ***
#> cor9     carb -0.5509251 1.084446e-03    **
#> cor8     gear  0.4802848 5.400948e-03    **
#> cor5     qsec  0.4186840 1.708199e-02     *