Correlation Table

Creates a correlation table using numeric variables.

Usage

ryx(data, y, x)

Arguments

data

a data frame

y

a numeric variable

x

string, a list of numeric variables. Defaults to all variables in the dataset provided.

Value

dataframe with correlations of y variable with all x-variables, p-value, and level of significance for each x variable.

Examples

ryx(mtcars, "mpg")
#> Correlations of mpg with
#>  variable          r            p sigif
#>        wt -0.8676594 1.293959e-10   ***
#>       cyl -0.8521620 6.112687e-10   ***
#>      disp -0.8475514 9.380327e-10   ***
#>        hp -0.7761684 1.787835e-07   ***
#>      drat  0.6811719 1.776240e-05   ***
#>        vs  0.6640389 3.415937e-05   ***
#>        am  0.5998324 2.850207e-04   ***
#>      carb -0.5509251 1.084446e-03    **
#>      gear  0.4802848 5.400948e-03    **
#>      qsec  0.4186840 1.708199e-02     *
ryx(iris, "Sepal.Length")
#> Correlations of Sepal.Length with
#>      variable          r            p sigif
#>  Petal.Length  0.8717538 1.038667e-47   ***
#>   Petal.Width  0.8179411 2.325498e-37   ***
#>   Sepal.Width -0.1175698 1.518983e-01      
library(MASS)
ryx(Boston, "medv")
#> Correlations of medv with
#>  variable          r            p sigif
#>     lstat -0.7376627 5.081103e-88   ***
#>        rm  0.6953599 2.487229e-74   ***
#>   ptratio -0.5077867 1.609509e-34   ***
#>     indus -0.4837252 4.900260e-31   ***
#>       tax -0.4685359 5.637734e-29   ***
#>       nox -0.4273208 7.065042e-24   ***
#>      crim -0.3883046 1.173987e-19   ***
#>       age -0.3769546 1.569982e-18   ***
#>        zn  0.3604453 5.713584e-17   ***
#>     black  0.3334608 1.318113e-14   ***
#>       dis  0.2499287 1.206612e-08   ***