Calculate Correlations Between Selected Variable and Other Variables

Calculates the correlations between a selected numeric variable and other numeric variables in a dataset. Users can specify which numeric variables to include in the comparison. If no numeric variable is specified the function will compare the selected variable with all of the other variables in the data set. The function also evaluates the statistical significance of each correlation.

Usage

ryx(data, y, x)

Arguments

data

A data frame containing the dataset to be analyzed.

y

A character string specifying the variable (must be numeric) for correlation analysis.

x

An optional character string specifying the variable that the user wants to compare variable y to. If not provided, all numeric variables in the dataset except y will be used.

Value

The function returns an object of class ryx, which includes y,x , df which is a data frame with the variables of interest, variable is the column name of the variable, r is the correlation coefficient, p is the p-value for the correlation test and sigif is the significance level indicators.

Examples

# Example usage with the MASS package
library(MASS)
object <- ryx(Boston, y = "medv")

# Example usage with the mtcars dataset
library(datasets)
object2 <- ryx(mtcars, y = "mpg")